Change the f2s_test to work with MSVC
Also add some more debugging code to f2s (only enabled with -DDEBUG).

The issue seems to be that MSVC rounds up 7.038531E-26 to the next
higher floating point value, rather than to the next lower one (it lies
exactly halfway between two values). Workaround is to change the
constant to 7.0385309E-26, which is clearly inside the lower range.