Rewrite Java implementation to use 32-bit factors
This puts it in line with the C implementation fallback code. This may
be required in order to half the lookup table size, which may require
higher bit counts, and the previous implementation was limited to 4*31 =
124 bits.

This may also be a deadend. If the code becomes significantly slower in
a way that we cannot recover from, then we may have to undo this.