Only access even-numbered entries in the lookup table
Naively, this requires 65 bit. However, we can immediately divide by 10
since we know that the last digit is unnecessary, and that is the same
as dividing by 2 and then by 5. We do this by increasing the shift
distance by one, so it's essentially for free.

However, this is somewhat costly in terms of performance.