Implement round-even for the Java 32-bit implementation
There are two rounding cases; we already round even for the case where the
shortest value coincides with one of the bounds. However, we don't round even
if exact value is exactly half-way between two shortest values, i.e., if it
ends with an additional 5 digit, followed by zeroes.

Progress on #8.