Improve fallback performance
In the fallback case, instead of doing three 64x128 bit multiplications, we do one, and then compute the other two values from that. This works since we need to compute (4m+2,4m,4m-1-x)*multiplicand. This is significantly faster for the fallback case, although it does not translate into better performance if we have 64x64 bit intrinsics or a 128 bit type.