Check the gap between the bits we have and those we need
If there is at least one zero bit in the gap, then we technically don't need
to compute the second half of the product, but can compute the result right
away.

This is not a performance win in this case, but it might be in other cases.
I tried both computing a mask on-the-fly and using precomputed masks, but there
was no significant difference.