Use 32-bit arithmetic to compute integer logarithms
The integer logarithms used in double_pow5bits and in the computation of
q can be computed using 32-bit arithmetic for the exponents used in the
double- and single-precision versions of Ryu. (The exponent range for
double-precision is -1074-2 <= e <= 972-2.)

Using an approximation of the form n/2^m for the contants, an integer
division by a constant can be replaced by a single right-shift.

For the multiplication in double_pow5bits, GCC was previously generating
a slow 64-bit div instruction (in the case e != 0).

The performance improvements for MSVC (19.14.26433 for x64) are small
but measurable. For GCC (MinGW 7.2 x64) the performance improvements are
quite large. I didn't run the benchmarks on 32-bit platforms.