Use 32-bit instead of 64-bit shifts on x86
On MSVC this removes the last dependency from one of the 64-bit
arithmetic helper functions: __aullshr.

Since only the upper 32-bits of the result are needed, this replaces
a 64-bit by a 32-bit funnel shift.

This code path is currently only enabled for the 32-bit MSVC compiler.
It uses the same predefined macros as the digit generation loop in
d2s_buffered.