Use log10 from Hacker's Delight
This is slightly slower with my original benchmark, which is a bit
unrealistic since it converts the same number 1000 times, so it gets
perfect branch prediction.

As a demo, I changed the benchmark to pre-generate a list of 1000
random numbers and convert each of them for each of the outer loops.
In this scenario, the algorithm from Hacker's Delight is a tiny bit
faster. Note that the benchmark returns generally higher times,
probably due to the branch prediction misses). It is arguable whether
this benchmark is more or less representative of real programs.