Swap order of decimal digit gen and binary to decimal conv
As proposed by @mitchblank, this introduces a 'presentation layer' and a
'conversion layer', with the former depending on the latter. The presentation
layer decodes the floating point number and then passes it to the conversion
layer to get the decimal digits (only if necessary). This implies that it has
the sign bit, so the floating decimal does not have to contain it.

For some unfathomable reason, this seems to be a bit faster.