Style: Avoid elaborated-type-specifiers.
In C++, `struct X {};` can be referred to as simply `X`. Saying
`struct X` is known as an elaborated-type-specifier; while that also
works, it is rarely seen, more verbose, and occasionally problematic
(as typoing `struct Y` results in a forward declaration).

This can be avoided with the common C practice of
`typedef struct X {} X;`, which allows both C and C++ to say simply
`X`. (Of course, the self-typedef is unnecessary in C++, but this
localizes the unnecessary code to the struct definition.)