Rearrange the generic interface, add float/double conversion
I tried to split unpacking and conversion into separate steps, but could not
make it work. Ryu requires knowing the number of bits in the fixed-width
floating point input representation, as well as whether the ieee exponent is 0
in order to compute mmShift, which is supposed to be an internal implementation
detail.

I considered passing in mantissaBits and exponentBits, either through a struct
or as additional params, but it's just awkward.

IIRC, 80-bit float is special in that it has an explicit first bit (whereas
all other ieee formats have an implicit first bit), so the current interface
won't work there either. Yuk.