Replace 64-bit division-by-constant by multiplications.
All tested compilers (MSVC, clang, gcc) generate call to library
functions when compiling for x86.

This commit replaces these divisions by multiplications and shifts.
The multipliers and shift values were taken from the code generated
by clang-x64.

For MSVC x86 the performance improvements are ~36%.

Before (with umul128 PR):
    Average & Stddev Ryu  Average & Stddev Grisu3
64:  173.605   19.360      466.816  262.839
64:  173.305   19.192      465.669  260.817
64:  175.686   24.207      473.329  263.606
64:  173.167   18.369      465.500  261.199

After (with umul128 PR):
    Average & Stddev Ryu  Average & Stddev Grisu3
64:  127.486   18.023      472.485  269.153
64:  125.630   13.030      465.577  261.548
64:  125.487   12.639      465.840  262.628
64:  125.522   13.553      465.153  262.756

The changes are only enabled for MSVC x86 for now and only for the hot
path of the algorithm.