benchmark.cc: Print hexfloats in verbose mode.
First, this extracts generate_float() and generate_double().

That eliminates the `r` integers, so we need another way to print the
exact data in verbose mode. C99's hexfloat conversion specifiers are
easy to use. "%.6a" and "%.13a" print enough hexits for round-tripping
floats and doubles.

Finally, we can also simplify %lf to %f; the arguments are doubles
(and C11 says that the 'l' length modifier "has no effect on a
following a, A, e, E, f, F, g, or G conversion specifier").