Fix signed/unsigned mismatch warnings, part 1.
olength is uint32_t (within [1, 9] for f2s and [1, 17] for d2s).
Casting it to int32_t is value-preserving and avoids triggering
compiler warnings when calculating int32_t exp.

(Previously, due to C's usual arithmetic conversions, the calculation
was performed as unsigned and then converted to signed. Because exp can
be negative, this could notionally involve converting a huge unsigned
number to negative signed. All modern machines are two's complement so
this doesn't make a difference to the bits, but fixing the compiler
warning also makes the code easier to reason about.)