Fix another typo-bug in unused code.
mulShift() and mulShift2() perform a wide multiplication
followed by a right shift and a modulo.

The last case, for j within [192, 255], which shifts away
everything except for the s1high part of the wide mult,
was incorrectly performing a left shift.

This was especially easy to miss because the other cases need `<< 32`.

This didn't affect correctness because the branch isn't being taken
(currently `j <= 180` holds), but this should be fixed just in case.