Fix yet another typo-bug in unused code.
mulShift2() was incorrectly performing an extra mod1e9() in one case,
unlike every other case in mulShift() and mulShift2().

This didn't affect correctness; previously, the broken case wasn't
activated due to the possible range of j, and now mulShift2() is
completely unused.

(I'm fixing this in case the code is ever needed again.)