Improve d2fixed_buffered_n() perf for `e2 == -53`.
We're printing `m2 * 2^e2`, and this branch is for the integer part
(to the left of the decimal point).

m2 is the ieeeMantissa with the implicit bit restored (0 for subnormal,
1 for normal), so it's a 53-bit number (at most).

When `e2 == -53`, `m2 * 2^-53` has no integer part, so this branch
doesn't need to be taken. (The following code will write a '0' for
the integer part, as it already does for more negative exponents.)

This is a strict performance improvement for this exponent,
and now this test is consistent with d2exp_buffered_n().

As this affects only one exponent, it's unobservable in the benchmark.