Test all binary exponents.
This verifies that all codepaths for `e2` are exercised. I've written
this with large constant tables, and refactored the previous
"all powers of ten" tests, to avoid pathological compile times.

Additionally, I've removed a thinko from the first version of my
"all powers of ten" tests. I failed to notice that 9e-324 (the second
smallest subnormal) and 1e-323 are exactly identical. As the smallest
subnormal is already tested, I simply eliminated the 9e-324 case.