Use bitwise OR in umul128() to improve codegen.
When concatenating two 32-bit values, MSVC's x86 codegen
is affected by plus versus bitwise OR.

For plus, the codegen is:

```
add eax, DWORD PTR _b00$1$[esp+16]
adc edi, 0
```

That's an add followed by add-with-carry;
i.e. the compiler is emitting a general 64 + 32 addition.

For bitwise OR, the codegen is:

```
or  eax, DWORD PTR _b00$1$[esp+16]
```

Bitwise OR is both conceptually simpler and more efficient.