Improve non-uint128 mulShift() carry logic.
In 57b918f, I saw two additions with only one overflow check for
carrying, and added an assertion. Now that I understand what the code
is doing, the assertion can be removed - but the overflow check
actually needs to be changed.

This code is forming and adding double-width products, represented as
separate parts (high0 and low0, etc.). This means that we can take
advantage of the double-width product property, but we need to be
careful to keep track of what each variable represents. We're adding:

```
               high0  low0
+        high1 low1
+ high2  low2
===========================
  s1high s1low s0high s0low
```

Adding low1 to high0 generates s0high. This is taking a double-width
product (high1 low1) and adding a single digit (high0), which can't
overflow beyond high1. However, it can internally carry within the
double-width value (which happens behind the scenes for
uint128 + uint64). Here, we need to manually carry with c1, which
will be added to high1 (and that cannot overflow).

Then we take high1 + c1 and add low2, generating s1low. Again this is
taking a double-width product (high2 low2) and adding a single digit
(high1 + c1), which can't overflow beyond high2. Again it can
internally carry, so we need a single check.

Here is where the code must be fixed: we're adding two digits,
low2 + (high1 + c1), so we need to compare against one of the inputs.
Either low2 or (high1 + c1) would be fine, but comparing against high1
is incorrect. (I believe that the code was getting away with it due
to the specific ranges of the inputs.) Since low2 is already available,
I'm using that.

Note that I have also changed the computation of c1 for stylistic
consistency, although there is no correctness issue there. (In a long
chain of overlapping additions, we would want to consistently compare
against the low part.)

Thanks again to @abolz for explaining the double-width product property
to me: https://github.com/ulfjack/ryu/issues/94#issuecomment-458942435