Reduce POW10_SPLIT_2 table size
Table size before: 4445
Table size after:  3133

Performance before:
%f:  159.717  170.618     4020.136 5110.341
%e:   49.362   12.508      501.819  220.803

Performance after:
%f:  147.380  173.127     3934.790 4996.622
%e:   45.730    8.462      464.647  191.248

Also add the Java class I've been using to generate the tables.

It was slightly more complex than I originally thought, because we no
longer have table entries < MIN_BLOCK_2, which we were previously
falling back to. That is fortunately unnecessary - if we only want to
print fewer blocks than we have, then we can just print all zeroes and
be done with it.

Also note that this change removes non-zero entries from POW10_SPLIT_2.
This is possible for cases where the factors are so small that they are
shifted out completely regardless of the magnitude of the mantissa.
I.e., we compute (m << 8) * FACTOR >> 128; if FACTOR is so small that
even the maximum value for m returns zero, then we can drop that factor.

This fixes #105.