F2s: deallocate char* after check
I couldn't find a way to make the compiler automatically deallocate the memory. Returning std::string from a wrapper function doesn't work - in that case the return value is still a pointer. Maybe wrapping it in another layer of std::unique_ptr would work, but that'd still make the tests more verbose. Using a macro seems a little nicer overall.

Progress on #121.