Add s2f
This is an almost exact copy of the s2d implementation. Only the corresponding
constants and types have been changed.

It should be easy (although it may take a while) to verify this implementation
for all possible 9-digit inputs and possibly detect errors; while it is almost
impossible to test all 17-digit inputs for the double-precision implementation.