Fix shiftRight128 for s2d
When called from s2d, the shift distance may be smaller than 32, so remove the
special-case. This looses some performance for d2s on 32-bit platforms if
RYU_OPTIMIZE_SIZE is *not* defined; if this is a problem, we could add another
variant to be used only by d2s.

Also add test coverage to CI.

Fixes #167.